lab5/requirements.txt
praw
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
mysql-connector-python==8.0.33
pandas
numpy
requests>=2.31.0
statsmodels==0.14.2
scikit-learn==1.5.1
pyarrow==16.1.0
gensim
scipy==1.12.0
